notebooks/06-app/requirements.txt
numpy==1.25.0
pandas==2.0.3
XlsxWriter==3.1.2
pyarrow==12.0.1
google-cloud-bigquery==3.11.3
pandas-gbq==0.19.2
altair==5.0.1
streamlit==1.24.0
